refactor(app): ER: Split FailedStepNextStep
This was a single component that had both the step lister thing and a
forced two-column layout, which isn't really useful. Make
FailedStepNextStep its own component and use it in the TwoCol.
