+++ app/src/organisms/ErrorRecoveryFlows/shared/TwoColTextAndFailedStepNextStep.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react'
-import { useTranslation } from 'react-i18next'
 import { css } from 'styled-components'
 import {
   DIRECTION_COLUMN,
@@ -10,11 +9,9 @@ import {
 } from '@opentrons/components'
 
 import { RecoveryContentWrapper } from './RecoveryContentWrapper'
-import {
-  TwoColumn,
-  CategorizedStepContent,
-} from '../../../molecules/InterventionModal'
+import { TwoColumn } from '../../../molecules/InterventionModal'
 import { RecoveryFooterButtons } from './RecoveryFooterButtons'
+import { FailedStepNextStep } from './FailedStepNextStep'
 
 import type { RecoveryContentProps } from '../types'
 
@@ -24,50 +21,24 @@ type TwoColTextAndFailedStepNextStepProps = RecoveryContentProps & {
   primaryBtnCopy: string
   primaryBtnOnClick: () => void
   secondaryBtnOnClickOverride?: () => void
-  secondaryBtnOnClickCopyOverride?: string
 }
 
 /**
  * Left Column: Title + body text
  * Right column: FailedStepNextStep
  */
-export function TwoColTextAndFailedStepNextStep({
-  leftColBodyText,
-  leftColTitle,
-  primaryBtnCopy,
-  primaryBtnOnClick,
-  secondaryBtnOnClickOverride,
-  secondaryBtnOnClickCopyOverride,
-  routeUpdateActions,
-  failedCommand,
-  stepCounts,
-  commandsAfterFailedCommand,
-  protocolAnalysis,
-  robotType,
-}: TwoColTextAndFailedStepNextStepProps): JSX.Element | null {
+export function TwoColTextAndFailedStepNextStep(
+  props: TwoColTextAndFailedStepNextStepProps
+): JSX.Element | null {
+  const {
+    leftColBodyText,
+    leftColTitle,
+    primaryBtnCopy,
+    primaryBtnOnClick,
+    secondaryBtnOnClickOverride,
+    routeUpdateActions,
+  } = props
   const { goBackPrevStep } = routeUpdateActions
-  const { t } = useTranslation('error_recovery')
-  const nthStepAfter = (n: number): number | undefined =>
-    stepCounts.currentStepNumber == null
-      ? undefined
-      : stepCounts.currentStepNumber + n
-  const nthCommand = (n: number): typeof failedCommand =>
-    commandsAfterFailedCommand != null
-      ? n < commandsAfterFailedCommand.length
-        ? commandsAfterFailedCommand[n]
-        : null
-      : null
-
-  const commandsAfter = [nthCommand(0), nthCommand(1)] as const
-
-  const indexedCommandsAfter = [
-    commandsAfter[0] != null
-      ? { command: commandsAfter[0], index: nthStepAfter(1) }
-      : null,
-    commandsAfter[1] != null
-      ? { command: commandsAfter[1], index: nthStepAfter(2) }
-      : null,
-  ] as const
 
   return (
     <RecoveryContentWrapper>
@@ -77,7 +48,7 @@ export function TwoColTextAndFailedStepNextStep({
           css={css`
             gap=${SPACING.spacing16};
             @media ${RESPONSIVENESS.touchscreenMediaQuerySpecs} {
-               gap=${SPACING.spacing8}
+            gap=${SPACING.spacing8}
             }
           `}
         >
@@ -94,23 +65,7 @@ export function TwoColTextAndFailedStepNextStep({
             {leftColBodyText}
           </StyledText>
         </Flex>
-        <CategorizedStepContent
-          commandTextData={protocolAnalysis}
-          robotType={robotType}
-          topCategoryHeadline={t('failed_step')}
-          topCategory="failed"
-          topCategoryCommand={
-            failedCommand == null
-              ? null
-              : {
-                  command: failedCommand,
-                  index: stepCounts.currentStepNumber ?? undefined,
-                }
-          }
-          bottomCategoryHeadline={t('next_step')}
-          bottomCategory="future"
-          bottomCategoryCommands={indexedCommandsAfter}
-        />
+        <FailedStepNextStep {...props} />
       </TwoColumn>
       <RecoveryFooterButtons
         primaryBtnOnClick={primaryBtnOnClick}
